Moved dummy classes for sequencing related tests into a module to increase reusability between tests.
