+++ tests/pulses/__init__.py
@@ -1,3 +1,3 @@
 from pkgutil import extend_path
-__all__ = ['ConditionTests','InstructionsTests','ParameterTests','SequencerTests','SequenceTests']
+__all__ = ['ConditionTests','InstructionsTests','ParameterTests','SequencerTests','SequenceTests','SequencingDummies']
 __path__ = extend_path(__path__, __name__)
